Bug/anand89/main general fixes (#109)
* bruh

* WS Init Untested

* Conflicts:
*	source/main_module/main.c

* throttle map

* fixes

* stale checks

* more stale fix

* CAN fixes

* CAN priority tx queues

* turnoff bootloader build

* can priority queue tested

* Update can tx for bootloader (untested)

* remove imd from bms line

* bump stale thresh to 250% of period

---------

+++ common/plettenberg/plettenberg.h
@@ -86,12 +86,12 @@ typedef struct
     uint32_t      rx_timeout;                       // Dynamically set timeout to determine connection status
 
     // Motor outputs
-    uint16_t      voltage_x10;
-    uint16_t      current_x10;
-    uint16_t      curr_power_x10;                   // Last torque command percent output sent x10
+    uint32_t      voltage_x10;
+    uint32_t      current_x10;
+    uint32_t      curr_power_x10;                   // Last torque command percent output sent x10
     uint32_t      rpm;
-    uint8_t       controller_temp;
-    uint8_t       motor_temp;
+    uint32_t       controller_temp;
+    uint32_t       motor_temp;
 
     // Communications
     q_handle_t   *tx_queue;                         // FIFO for tx commands to be sent via DMA
